Move user profile settings to modal
