+++ app/user-profile/user-profile-module.js
@@ -3,7 +3,8 @@ require('angular-xeditable');
 angular.module('ushahidi.user-profile', [
     'xeditable'
 ])
+.directive('accountSettings', require('./directives/account-settings.js'))
+.directive('userProfile', require('./directives/user-profile.js'))
+.directive('notifications', require('./directives/notifications.js'))
 .service('ContactEndpoint', require('./services/endpoints/contact.js'))
-.service('NotificationEndpoint', require('./services/endpoints/notification.js'))
-.controller('userProfile', require('./controllers/user-profile-controller.js'))
-.config(require('./user-profile-routes.js'));
+.service('NotificationEndpoint', require('./services/endpoints/notification.js'));
