--- app/user-profile/user-profile-routes.js
@@ -1,23 +1,0 @@
-module.exports = [
-    '$routeProvider',
-function (
-    $routeProvider
-) {
-
-
-    $routeProvider
-    .when('/users/me', {
-        controller: require('./controllers/user-profile-controller.js'),
-        templateUrl: 'templates/users/me.html',
-        resolve: {
-            user: ['UserEndpoint', function (UserEndpoint) {
-                return UserEndpoint.getFresh({id: 'me'}).$promise;
-            }]
-        }
-    })
-    .when('/users/me/notifications/', {
-        controller: require('./controllers/notification-controller.js'),
-        templateUrl: 'templates/notifications/notifications.html'
-    })
-    ;
-}];
