+++ server/www/index.html
@@ -42,7 +42,7 @@
         <div id="bootstrap-app" class="hidden">
 
             <mode-bar></mode-bar>
-
+            <account-settings></account-settings>
             <ng-view></ng-view>
 
             <!-- TODO: fix and uncomment -->
