+++ server/www/templates/frame/mode-bar.html
@@ -56,7 +56,7 @@
                 </a>
             </li>
             <li ng-show="currentUser">
-                <a>
+                <a ng-click="viewAccountSettings()">
                     <img class="avatar" src="http://www.gravatar.com/avatar/{{ currentUser.gravatar || '00000000000000000000000000000000' }}?d=retro" alt="{{ currentUser.realname }}"/>
                     <span class="label">{{currentUser.realname || currentUser.email}}</span>
                 </a>
